+++ material/src/modules/confirmationDialog.module.ts
@@ -3,6 +3,7 @@ import {NgModule} from '@angular/core';
 import {ConfirmationDialogComponent} from '../components/confirmationDialog/confirmationDialog.component';
 import {TitledDialogModule} from './titledDialog.module';
 import {ConfirmationDialogDirective} from '../directives/confirmationDialog/confirmationDialog.directive';
+import {ConfirmationDialogChoiceTemplateDirective} from '../directives/confirmationDialogChoiceTemplate/confirmationDialogChoiceTemplate.directive';
 
 /**
  * Module containing confirmation dialog component and directive
@@ -13,11 +14,13 @@ import {ConfirmationDialogDirective} from '../directives/confirmationDialog/conf
     [
         TitledDialogModule,
         ConfirmationDialogDirective,
+        ConfirmationDialogChoiceTemplateDirective,
         ConfirmationDialogComponent,
     ],
     exports:
     [
         ConfirmationDialogDirective,
+        ConfirmationDialogChoiceTemplateDirective,
     ]
 })
 export class ConfirmationDialogModule